Added new token, refined tests
