+++ packages/pn-pa-webapp/cypress/e2e/Auth.cy.ts
@@ -6,15 +6,13 @@ describe('Auth', () => {
     cy.viewport(1920, 1080);
   });
 
-  it.skip('Should login properly through UI', () => {
-    cy.logout();
-    cy.login();
-    cy.loginWithUI();
-  });
-
   it('Should login properly using APIs', () => {
     cy.logout();
-    cy.login();
+    cy.loginWithTokenExchange();
+    cy.window().then(win => {
+      const user = win.sessionStorage.getItem('user');
+      expect(user).not.to.be.null;
+    });
   });
 
   it('Should logout properly using APIs', () => {
